settings: Move links into Links component
