+++ src/components/SettingsManager/SettingsPanel.js
@@ -2,33 +2,15 @@ import cx from 'classnames';
 import React from 'react';
 import PropTypes from 'prop-types';
 import { translate } from 'react-i18next';
-import FlatButton from 'material-ui/FlatButton';
-import LicenseIcon from 'material-ui/svg-icons/action/copyright';
-import GithubIcon from './GithubIcon';
 import Profile from './Profile';
 import LabeledControl from './LabeledControl';
 import LanguagePicker from './LanguagePicker';
 import LogoutButton from './LogoutButton';
 import NotificationSettings from './NotificationSettings';
+import Links from './Links';
 import Toggle from './Toggle';
 
-const linkStyle = {
-  display: 'block',
-  height: 24,
-  lineHeight: '24px',
-  marginBottom: 20,
-  textAlign: 'left',
-  WebkitAppearance: 'initial'
-};
-const iconStyle = { verticalAlign: 'top' };
-
-const linkProps = {
-  style: linkStyle,
-  target: '_blank',
-  labelPosition: 'after',
-  backgroundColor: 'transparent',
-  hoverColor: 'transparent'
-};
+const enhance = translate();
 
 class SettingsPanel extends React.Component {
   static propTypes = {
@@ -68,18 +50,15 @@ class SettingsPanel extends React.Component {
       onLogout
     } = this.props;
 
-    const profile = user && [
-      <Profile
-        key="profile"
-        user={user}
-        onChangeUsername={onChangeUsername}
-      />,
-      <hr key="divider" className="SettingsPanel-divider" />
-    ];
-
     return (
       <div className={cx('SettingsPanel', className)}>
-        {profile}
+        {user && (
+          <Profile
+            user={user}
+            onChangeUsername={onChangeUsername}
+          />
+        )}
+        {user && <hr className="SettingsPanel-divider" />}
         <div className="SettingsPanel-column SettingsPanel-column--left">
           <h2 className="SettingsPanel-header">{t('settings.title')}</h2>
           <Toggle
@@ -113,32 +92,12 @@ class SettingsPanel extends React.Component {
             settings={settings}
             onSettingChange={this.props.onSettingChange}
           />
-          <h2 className="SettingsPanel-header">{t('settings.links.title')}</h2>
-          <FlatButton
-            href="https://github.com/u-wave"
-            label={t('settings.links.website')}
-            {...linkProps}
-          >
-            <GithubIcon style={iconStyle} />
-          </FlatButton>
-          <FlatButton
-            href="https://github.com/u-wave/web"
-            label={t('settings.links.source')}
-            {...linkProps}
-          >
-            <GithubIcon style={iconStyle} />
-          </FlatButton>
-          <FlatButton
-            href="https://github.com/u-wave/web/tree/master/LICENSE"
-            label={t('settings.links.license')}
-            {...linkProps}
-          >
-            <LicenseIcon style={iconStyle} />
-          </FlatButton>
+          <hr className="SettingsPanel-divider" />
+          <Links />
         </div>
       </div>
     );
   }
 }
 
-export default translate()(SettingsPanel);
+export default enhance(SettingsPanel);
